делаю PointController и переношу туда логику по отображению карточек
